refactor: componentsをleftPane/rightPane/sharedに整理
ResultTable.tsを機能単位で分割し、コンポーネントディレクトリを
leftPane（入力・設定系）、rightPane（結果表示系）、shared（共通ユーティリティ）に再構成。

- leftPane: FileInput, CrossConfig, SavedFiles
- rightPane: ResultView（オーケストレーター）, Toolbar, GtTable, CrossTable, ChartRenderer, AIBubble
- shared: escHtml（重複していたHTML エスケープ関数を共通化）
