+++ src/components/shared/escHtml.ts
@@ -1,0 +1,3 @@
+export function escHtml(str: string): string {
+  return String(str).replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
+}
